+++ presentation/src/main/java/com/aga/presentation/base/Constants.kt
@@ -1,4 +1,6 @@
 package com.aga.presentation.base
 
 object Constants {
+
+    const val LOGIN_TO_JOIN = 10001
 }
